apps/agent/pyproject.toml
[project]
name = "cavaticus-agent"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "anthropic>=0.45.0",
    "openai>=1.65.0",
    "google-genai>=1.5.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "tinycss2>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
